+++ arg_picker/src/builtin.rs
@@ -3,6 +3,7 @@ mod pick_flag;
 mod pick_ip_attr;
 mod pick_numbers;
 mod pick_pathbuf;
+mod pick_paths;
 mod pick_picker_args;
 mod pick_socket_attr;
 mod pick_string;