+++ arg_picker/src/value.rs
@@ -1,5 +1,8 @@
 mod flag;
 pub use flag::*;
 
+mod paths;
+pub use paths::*;
+
 mod vec_until;
 pub use vec_until::*;